More self cleanup
