+++ pkg/relayer/cmd/keys.go
@@ -8,13 +8,12 @@ import (
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	cosmostypes "github.com/cosmos/cosmos-sdk/types"
+
+	relayerconfig "github.com/pokt-network/poktroll/pkg/relayer/config"
 )
 
 // getPrivateKeyHexFromKeyring takes a key name and returns the private key in hex format
-// func getPrivateKeyHexFromKeyring(kr keyring.Keyring, keyName string) (string, error) {
 func getPrivateKeyHexFromKeyring(kr keyring.Keyring, address string) (string, error) {
-	// Export the private key in armored format
-	// armoredPrivKey, err := kr.ExportPrivKeyArmor(keyName, "") // Empty passphrase
 	cosmosAddr := cosmostypes.MustAccAddressFromBech32(address)
 	armoredPrivKey, err := kr.ExportPrivKeyArmorByAddress(cosmosAddr, "") // Empty passphrase
 	if err != nil {
@@ -36,4 +35,23 @@ func getPrivateKeyHexFromKeyring(kr keyring.Keyring, address string) (string, er
 	// Convert to hex
 	hexKey := hex.EncodeToString(secpPrivKey.Key)
 	return hexKey, nil
+}
+
+// uniqueSigningKeyNames returns a list of unique operator signing key names from the RelayMiner config.
+func uniqueSigningKeyNames(relayMinerConfig *relayerconfig.RelayMinerConfig) []string {
+	uniqueKeyMap := make(map[string]bool)
+	for _, server := range relayMinerConfig.Servers {
+		for _, supplier := range server.SupplierConfigsMap {
+			for _, signingKeyName := range supplier.SigningKeyNames {
+				uniqueKeyMap[signingKeyName] = true
+			}
+		}
+	}
+
+	uniqueKeyNames := make([]string, 0, len(uniqueKeyMap))
+	for key := range uniqueKeyMap {
+		uniqueKeyNames = append(uniqueKeyNames, key)
+	}
+
+	return uniqueKeyNames
 }
